worker/requirements.txt
celery[redis]
mysql-connector
